+++ packages/open-truss/src/configuration/components.tsx
@@ -1,0 +1,68 @@
+import type React from 'react'
+import { z } from 'zod'
+import {
+  type COMPONENTS,
+  type OpenTrussComponent,
+  type OpenTrussComponentExports,
+} from './apply'
+import { DataV1Shape, WorkflowV1Shape } from './workflow-config'
+
+export const BaseOpenTrussComponentV1PropsShape = z.object({
+  data: DataV1Shape,
+  config: WorkflowV1Shape,
+})
+export const withChildren = (shape: z.AnyZodObject): z.AnyZodObject =>
+  shape.extend({ children: z.any().optional() })
+const ComponentWithChildrenShape = withChildren(
+  BaseOpenTrussComponentV1PropsShape,
+)
+type ComponentWithChildren = z.infer<typeof ComponentWithChildrenShape>
+
+export type BaseOpenTrussComponentV1Props =
+  | z.infer<typeof BaseOpenTrussComponentV1PropsShape>
+  | ComponentWithChildren
+
+export type BaseOpenTrussComponentV1 = (
+  props: BaseOpenTrussComponentV1Props,
+) => React.JSX.Element
+
+function hasDefaultExport(
+  component: any,
+): component is OpenTrussComponentExports {
+  return 'default' in component
+}
+
+function hasPropsExport(
+  component: any,
+): component is OpenTrussComponentExports {
+  return 'Props' in component
+}
+
+export function hasChildren(
+  component: any,
+): component is ComponentWithChildren {
+  if (hasPropsExport(component)) {
+    return 'children' in component.Props.shape
+  }
+  // Default to true for legacy component definitions
+  return true
+}
+
+export function getComponent(
+  component: string,
+  COMPONENTS: COMPONENTS,
+): OpenTrussComponent {
+  let Component = COMPONENTS[component]
+  if (!Component) {
+    throw new Error(`No component '${component}' configured.`)
+  }
+  if (hasDefaultExport(Component)) {
+    Component = Component.default
+  }
+
+  if (!Component) {
+    throw new Error(`No component '${component}' configured.`)
+  }
+
+  return Component
+}
