+++ packages/open-truss/src/configuration/engine-v1.tsx
@@ -1,85 +1,9 @@
 import React from 'react'
-import { z } from 'zod'
-import { YamlObjectShape, YamlShape } from '../utils/yaml'
 import DataProvider from './DataProvider'
-import {
-  type COMPONENTS,
-  type OpenTrussComponent,
-  type OpenTrussComponentExports,
-  type ReactTree,
-  type RenderingEngine,
-} from './apply'
+import { type COMPONENTS, type ReactTree, type RenderingEngine } from './apply'
 import { processProps } from './process-props'
-
-const DataV1Shape = YamlShape.optional()
-export type DataV1 = z.infer<typeof DataV1Shape>
-const ViewPropsV1Shape = YamlObjectShape.optional()
-export type ViewPropsV1 = z.infer<typeof ViewPropsV1Shape>
-const FrameBase = z.object({
-  frame: z.null(), // used only to make configs more readable
-  view: z.object({
-    component: z.string(),
-    props: ViewPropsV1Shape,
-  }),
-  data: DataV1Shape,
-})
-
-export type FrameType = z.infer<typeof FrameBase> & {
-  frames?: FrameType[]
-}
-
-const FrameV1Shape: z.ZodType<FrameType> = FrameBase.extend({
-  frames: z
-    .lazy(() => FrameV1Shape)
-    .array()
-    .optional(),
-})
-type FrameV1 = z.infer<typeof FrameV1Shape>
-
-const WorkflowV1Shape = z.object({
-  version: z.number(),
-  frames: FrameV1Shape.array(),
-})
-export type WorkflowV1 = z.infer<typeof WorkflowV1Shape>
-
-export const BaseOpenTrussComponentV1PropsShape = z.object({
-  data: DataV1Shape,
-  config: WorkflowV1Shape,
-})
-export const withChildren = (shape: z.AnyZodObject): z.AnyZodObject =>
-  shape.extend({ children: z.any().optional() })
-const ComponentWithChildrenShape = withChildren(
-  BaseOpenTrussComponentV1PropsShape,
-)
-type ComponentWithChildren = z.infer<typeof ComponentWithChildrenShape>
-
-export type BaseOpenTrussComponentV1Props =
-  | z.infer<typeof BaseOpenTrussComponentV1PropsShape>
-  | ComponentWithChildren
-
-export type BaseOpenTrussComponentV1 = (
-  props: BaseOpenTrussComponentV1Props,
-) => React.JSX.Element
-
-function hasDefaultExport(
-  component: any,
-): component is OpenTrussComponentExports {
-  return 'default' in component
-}
-
-function hasPropsExport(
-  component: any,
-): component is OpenTrussComponentExports {
-  return 'Props' in component
-}
-
-function hasChildren(component: any): component is ComponentWithChildren {
-  if (hasPropsExport(component)) {
-    return 'children' in component.Props.shape
-  }
-  // Default to true for legacy component definitions
-  return true
-}
+import { type WorkflowV1, type FrameV1 } from './workflow-config'
+import { getComponent, hasChildren } from './components'
 
 export function engineV1(
   COMPONENTS: COMPONENTS,
@@ -122,23 +46,4 @@ export function engineV1(
   return () => {
     return renderFrames(config.frames)
   }
-}
-
-export function getComponent(
-  component: string,
-  COMPONENTS: COMPONENTS,
-): OpenTrussComponent {
-  let Component = COMPONENTS[component]
-  if (!Component) {
-    throw new Error(`No component '${component}' configured.`)
-  }
-  if (hasDefaultExport(Component)) {
-    Component = Component.default
-  }
-
-  if (!Component) {
-    throw new Error(`No component '${component}' configured.`)
-  }
-
-  return Component
 }
