+++ packages/open-truss/src/configuration/index.ts
@@ -1,2 +1,3 @@
 export * from './apply'
 export * from './engine-v1'
+export { BaseOpenTrussComponentV1PropsShape, withChildren } from './components'
